requirements.txt
# Processamento de Dados
pandas
numpy
openpyxl

# Dashboard e Visualização
streamlit
plotly
Pillow

# Outros (Dependências implícitas)
datetime